+++ src/components/dialog/Dialog.js
@@ -7,15 +7,25 @@ import theme from './theme.css';
 
 import { Button, ButtonGroup, DialogBase, Heading2, Heading3, Link } from '../../index';
 import { COLORS } from '../../constants';
+import { IconDragMediumFilled } from '@teamleader/ui-icons';
 
 class Dialog extends PureComponent {
   bodyRef = createRef();
+  dragHandleRef = createRef();
 
   getHeader = () => {
-    const { headerColor, headerIcon, headingLevel, onCloseClick, title } = this.props;
+    const { draggable, headerColor, headerIcon, headingLevel, onCloseClick, title } = this.props;
+
+    const icon = draggable ? (
+      <div className={theme['drag-icon']} ref={this.dragHandleRef}>
+        <IconDragMediumFilled />
+      </div>
+    ) : (
+      headerIcon
+    );
 
     return (
-      <DialogBase.Header color={headerColor} icon={headerIcon} onClose={onCloseClick}>
+      <DialogBase.Header color={headerColor} icon={icon} onClose={onCloseClick}>
         {headingLevel === 2 ? <Heading2>{title}</Heading2> : <Heading3>{title}</Heading3>}
       </DialogBase.Header>
     );
@@ -49,7 +59,13 @@ class Dialog extends PureComponent {
     ]);
 
     return (
-      <DialogBase className={classNames} {...restProps} scrollable={false} initialFocusRef={this.bodyRef}>
+      <DialogBase
+        className={classNames}
+        {...restProps}
+        scrollable={false}
+        initialFocusRef={this.bodyRef}
+        dragHandleRef={this.dragHandleRef}
+      >
         {title && this.getHeader()}
         <DialogBase.Body ref={this.bodyRef} scrollable={scrollable}>
           {children}
@@ -87,6 +103,8 @@ Dialog.propTypes = {
   tertiaryAction: PropTypes.object,
   /** The title of the dialog. */
   title: PropTypes.string,
+  /** If true, the dialog will be draggable. */
+  draggable: PropTypes.bool,
 };
 
 Dialog.defaultProps = {